Created new page drop down selector for type of taper setup taper frequncy and inputs works well
